[v5] Remove dead MSTestDiscoverer / MSTestExecutor VSTest wrappers
After the VSTest adapter surface was removed, these two classes were only thin
VSTest-shaped wrappers (unwrapping IDiscoveryContext / IRunContext /
IFrameworkHandle) that delegated to the neutral MSTestEngine. The native
Microsoft.Testing.Platform framework (MSTestTestFramework) already drives
MSTestEngine directly, so nothing in production instantiated them anymore --
only their own unit tests did.

- Delete MSTestDiscoverer.cs and MSTestExecutor.cs.
- Relocate the [ModuleInitializer] adapter/framework version check
  (EnsureAdapterAndFrameworkVersions) from MSTestExecutor into MSTestEngine,
  preserving the production behavior. Update InternalAPI Shipped/Unshipped
  accordingly.
- Delete MSTestExecutorTests and the MSTestDiscoverer-specific tests; keep the
  MSTestDiscovererHelpers / AdapterSettingsException coverage in the renamed
  MSTestDiscovererHelpersTests.

Copilot-Session: cf0134cb-060e-4d91-97cc-508708cad967
