+++ src/Adapter/MSTest.TestAdapter/InternalAPI/uwp/InternalAPI.Shipped.txt
@@ -7,16 +7,6 @@ Microsoft.VisualStudio.TestPlatform.MSTest.TestAdapter.Extensions.TestCaseExtens
 Microsoft.VisualStudio.TestPlatform.MSTest.TestAdapter.Extensions.TestResultExtensions
 Microsoft.VisualStudio.TestPlatform.MSTest.TestAdapter.Extensions.UnitTestElementExtensions
 Microsoft.VisualStudio.TestPlatform.MSTest.TestAdapter.Helpers.UnitTestOutcomeHelper
-Microsoft.VisualStudio.TestPlatform.MSTest.TestAdapter.MSTestDiscoverer
-Microsoft.VisualStudio.TestPlatform.MSTest.TestAdapter.MSTestDiscoverer.DiscoverTestsAsync(System.Collections.Generic.IEnumerable<string!>! sources, Microsoft.VisualStudio.TestPlatform.ObjectModel.Adapter.IDiscoveryContext! discoveryContext, Microsoft.VisualStudio.TestPlatform.ObjectModel.Logging.IMessageLogger! logger, Microsoft.VisualStudio.TestPlatform.MSTestAdapter.PlatformServices.Interface.IUnitTestElementSink! elementSink, Microsoft.VisualStudio.TestPlatform.MSTestAdapter.PlatformServices.Interface.IConfiguration? configuration, bool isMTP) -> System.Threading.Tasks.Task!
-Microsoft.VisualStudio.TestPlatform.MSTest.TestAdapter.MSTestDiscoverer.DiscoverTestsAsync(System.Collections.Generic.IEnumerable<string!>! sources, Microsoft.VisualStudio.TestPlatform.ObjectModel.Adapter.IDiscoveryContext! discoveryContext, Microsoft.VisualStudio.TestPlatform.ObjectModel.Logging.IMessageLogger! logger, Microsoft.VisualStudio.TestPlatform.ObjectModel.Adapter.ITestCaseDiscoverySink! discoverySink, Microsoft.VisualStudio.TestPlatform.MSTestAdapter.PlatformServices.Interface.IConfiguration? configuration, bool isMTP) -> System.Threading.Tasks.Task!
-Microsoft.VisualStudio.TestPlatform.MSTest.TestAdapter.MSTestDiscoverer.MSTestDiscoverer(Microsoft.VisualStudio.TestPlatform.MSTestAdapter.PlatformServices.Interface.ITestSourceHandler! testSourceHandler, System.Func<string!, System.Collections.Generic.IDictionary<string!, object!>!, System.Threading.Tasks.Task!>? telemetrySender = null) -> void
-Microsoft.VisualStudio.TestPlatform.MSTest.TestAdapter.MSTestExecutor
-Microsoft.VisualStudio.TestPlatform.MSTest.TestAdapter.MSTestExecutor.Cancel() -> void
-Microsoft.VisualStudio.TestPlatform.MSTest.TestAdapter.MSTestExecutor.MSTestExecutor(System.Threading.CancellationToken cancellationToken, System.Func<string!, System.Collections.Generic.IDictionary<string!, object!>!, System.Threading.Tasks.Task!>? telemetrySender = null) -> void
-Microsoft.VisualStudio.TestPlatform.MSTest.TestAdapter.MSTestExecutor.RunTestsAsync(System.Collections.Generic.IEnumerable<string!>? sources, Microsoft.VisualStudio.TestPlatform.ObjectModel.Adapter.IRunContext? runContext, Microsoft.VisualStudio.TestPlatform.ObjectModel.Adapter.IFrameworkHandle? frameworkHandle, Microsoft.VisualStudio.TestPlatform.MSTestAdapter.PlatformServices.Interface.IConfiguration? configuration, bool isMTP) -> System.Threading.Tasks.Task!
-Microsoft.VisualStudio.TestPlatform.MSTest.TestAdapter.MSTestExecutor.RunTestsAsync(System.Collections.Generic.IEnumerable<string!>? sources, Microsoft.VisualStudio.TestPlatform.ObjectModel.Adapter.IRunContext? runContext, Microsoft.VisualStudio.TestPlatform.ObjectModel.Adapter.IFrameworkHandle? frameworkHandle, System.Func<Microsoft.VisualStudio.TestPlatform.MSTest.TestAdapter.MSTestSettings!, Microsoft.VisualStudio.TestPlatform.MSTestAdapter.PlatformServices.Interface.ITestResultRecorder!>! recorderFactory, Microsoft.VisualStudio.TestPlatform.MSTestAdapter.PlatformServices.Interface.IConfiguration? configuration, bool isMTP) -> System.Threading.Tasks.Task!
-Microsoft.VisualStudio.TestPlatform.MSTest.TestAdapter.MSTestExecutor.TestExecutionManager.get -> Microsoft.VisualStudio.TestPlatform.MSTest.TestAdapter.Execution.TestExecutionManager!
 Microsoft.VisualStudio.TestPlatform.MSTest.TestAdapter.TestElementFilterProvider
 Microsoft.VisualStudio.TestPlatform.MSTest.TestAdapter.TestElementFilterProvider.GetTestElementFilter(Microsoft.VisualStudio.TestPlatform.MSTestAdapter.PlatformServices.Interface.IAdapterMessageLogger! logger, out bool filterHasError) -> Microsoft.VisualStudio.TestPlatform.MSTestAdapter.PlatformServices.Interface.ITestElementFilter?
 Microsoft.VisualStudio.TestPlatform.MSTest.TestAdapter.TestElementFilterProvider.TestElementFilterProvider(Microsoft.VisualStudio.TestPlatform.ObjectModel.Adapter.IDiscoveryContext? context) -> void
@@ -48,7 +38,6 @@ static Microsoft.VisualStudio.TestPlatform.MSTest.TestAdapter.Extensions.UnitTes
 static Microsoft.VisualStudio.TestPlatform.MSTest.TestAdapter.Extensions.UnitTestElementExtensions.ToTestCase(this Microsoft.VisualStudio.TestPlatform.MSTest.TestAdapter.ObjectModel.UnitTestElement! element) -> Microsoft.VisualStudio.TestPlatform.ObjectModel.TestCase!
 static Microsoft.VisualStudio.TestPlatform.MSTest.TestAdapter.Extensions.UnitTestElementExtensions.VersionedGuidFromHash(byte[]! hashBytes, byte hashVersion) -> System.Guid
 static Microsoft.VisualStudio.TestPlatform.MSTest.TestAdapter.Helpers.UnitTestOutcomeHelper.ToTestOutcome(Microsoft.VisualStudio.TestTools.UnitTesting.UnitTestOutcome unitTestOutcome, Microsoft.VisualStudio.TestPlatform.MSTest.TestAdapter.MSTestSettings! currentSettings) -> Microsoft.VisualStudio.TestPlatform.ObjectModel.TestOutcome
-static Microsoft.VisualStudio.TestPlatform.MSTest.TestAdapter.MSTestExecutor.MSTestModuleInitializer() -> void
 static Microsoft.VisualStudio.TestPlatform.MSTest.TestAdapter.TestResultRecorderExtensions.ToTestResultRecorder(this Microsoft.VisualStudio.TestPlatform.ObjectModel.Adapter.ITestExecutionRecorder! testExecutionRecorder, string! computerName, Microsoft.VisualStudio.TestPlatform.MSTest.TestAdapter.MSTestSettings! settings) -> Microsoft.VisualStudio.TestPlatform.MSTestAdapter.PlatformServices.Interface.ITestResultRecorder!
 static Microsoft.VisualStudio.TestPlatform.MSTestAdapter.PlatformServices.AdapterMessageLoggerExtensions.ToAdapterMessageLogger(this Microsoft.VisualStudio.TestPlatform.ObjectModel.Logging.IMessageLogger! messageLogger) -> Microsoft.VisualStudio.TestPlatform.MSTestAdapter.PlatformServices.Interface.IAdapterMessageLogger!
 static Microsoft.VisualStudio.TestPlatform.MSTestAdapter.PlatformServices.MessageLevelExtensions.ToTestMessageLevel(this Microsoft.VisualStudio.TestTools.UnitTesting.MessageLevel messageLevel) -> Microsoft.VisualStudio.TestPlatform.ObjectModel.Logging.TestMessageLevel
