--- src/Adapter/MSTest.TestAdapter/VSTestAdapter/MSTestExecutor.cs
@@ -1,102 +1,0 @@
-﻿// Copyright (c) Microsoft Corporation. All rights reserved.
-// Licensed under the MIT license. See LICENSE file in the project root for full license information.
-
-using Microsoft.VisualStudio.TestPlatform.MSTest.TestAdapter.Execution;
-using Microsoft.VisualStudio.TestPlatform.MSTestAdapter.PlatformServices;
-using Microsoft.VisualStudio.TestPlatform.MSTestAdapter.PlatformServices.Interface;
-using Microsoft.VisualStudio.TestPlatform.MSTestAdapter.PlatformServices.Resources;
-using Microsoft.VisualStudio.TestPlatform.ObjectModel.Adapter;
-using Microsoft.VisualStudio.TestTools.UnitTesting;
-
-namespace Microsoft.VisualStudio.TestPlatform.MSTest.TestAdapter;
-
-/// <summary>
-/// Contains the execution logic for this adapter.
-/// </summary>
-[StackTraceHidden]
-internal sealed class MSTestExecutor
-{
-    private readonly MSTestEngine _engine;
-
-    internal MSTestExecutor(CancellationToken cancellationToken, Func<string, IDictionary<string, object>, Task>? telemetrySender = null)
-    {
-        TestExecutionManager = new TestExecutionManager();
-        _engine = new MSTestEngine(cancellationToken, telemetrySender, TestExecutionManager);
-    }
-
-    /// <summary>
-    /// Gets the ms test execution manager.
-    /// </summary>
-    internal TestExecutionManager TestExecutionManager { get; }
-
-#pragma warning disable CA2255 // The 'ModuleInitializer' attribute should not be used in libraries
-    [ModuleInitializer]
-#pragma warning restore CA2255 // The 'ModuleInitializer' attribute should not be used in libraries
-    internal static void MSTestModuleInitializer()
-        => EnsureAdapterAndFrameworkVersions();
-
-    private static void EnsureAdapterAndFrameworkVersions()
-    {
-        string? adapterVersion = typeof(MSTestExecutor).Assembly.GetCustomAttribute<AssemblyInformationalVersionAttribute>()?.InformationalVersion;
-        string? frameworkVersion = typeof(TestMethodAttribute).Assembly.GetCustomAttribute<AssemblyInformationalVersionAttribute>()?.InformationalVersion;
-        if (adapterVersion is not null && frameworkVersion is not null
-            && adapterVersion != frameworkVersion)
-        {
-            throw new InvalidOperationException(string.Format(CultureInfo.InvariantCulture, Resource.VersionMismatchBetweenAdapterAndFramework, adapterVersion, frameworkVersion));
-        }
-    }
-
-    internal Task RunTestsAsync(IEnumerable<string>? sources, IRunContext? runContext, IFrameworkHandle? frameworkHandle, IConfiguration? configuration, bool isMTP)
-        => RunTestsFromSourcesCoreAsync(sources, runContext, frameworkHandle, settings => frameworkHandle!.ToTestResultRecorder(Environment.MachineName, settings), configuration, isMTP);
-
-    /// <summary>
-    /// Runs the tests from sources, reporting results to a caller-provided platform-agnostic
-    /// <see cref="ITestResultRecorder"/> (used by the native Microsoft.Testing.Platform integration, which
-    /// publishes test nodes itself instead of recording through the VSTest <see cref="IFrameworkHandle"/>).
-    /// The <paramref name="frameworkHandle"/> is still used for message logging and apartment-state handling.
-    /// The recorder is created via <paramref name="recorderFactory"/> after settings are resolved so it observes
-    /// the effective <see cref="MSTestSettings"/>.
-    /// </summary>
-    internal Task RunTestsAsync(IEnumerable<string>? sources, IRunContext? runContext, IFrameworkHandle? frameworkHandle, Func<MSTestSettings, ITestResultRecorder> recorderFactory, IConfiguration? configuration, bool isMTP)
-        => RunTestsFromSourcesCoreAsync(sources, runContext, frameworkHandle, recorderFactory, configuration, isMTP);
-
-    private async Task RunTestsFromSourcesCoreAsync(IEnumerable<string>? sources, IRunContext? runContext, IFrameworkHandle? frameworkHandle, Func<MSTestSettings, ITestResultRecorder> recorderFactory, IConfiguration? configuration, bool isMTP)
-    {
-        if (PlatformServiceProvider.Instance.AdapterTraceLogger.IsInfoEnabled)
-        {
-            PlatformServiceProvider.Instance.AdapterTraceLogger.Info("MSTestExecutor.RunTests: Running tests from sources.");
-        }
-
-        if (frameworkHandle is null)
-        {
-            throw new ArgumentNullException(nameof(frameworkHandle));
-        }
-
-        // VSTest annotates the IEnumerable as nullable; the exact reason is unclear.
-        if (sources is null)
-        {
-            throw new ArgumentNullException(nameof(sources));
-        }
-
-        // Unwrap the VSTest run context / framework handle into neutral inputs and delegate to the
-        // platform-agnostic engine. The recorder is created via the factory after settings are resolved so it
-        // observes the effective MSTestSettings; for the VSTest path this wraps the framework handle, for the
-        // native MTP path it publishes test nodes directly.
-        await _engine.RunFromSourcesAsync(
-            sources,
-            runContext?.RunSettings?.SettingsXml,
-            runContext?.TestRunDirectory,
-            frameworkHandle.ToAdapterMessageLogger(),
-            recorderFactory,
-            new TestElementFilterProvider(runContext),
-            configuration,
-            new TestSourceHandler(),
-            isMTP).ConfigureAwait(false);
-    }
-
-    /// <summary>
-    /// Cancel the test run.
-    /// </summary>
-    internal void Cancel()
-        => _engine.Cancel();
-}
